refactor(gui): begin moving signal handlers out of main
For testing, signal handlers will all be moved out of the main function
and into their own functions. It's annoying to create an Application and
MainWindow, so these functions will send to a channel with the
information needed to make the necessary changes to the MainWindow. This
way the state changes are decoupled from the window itself.

This commit only moves refresh_devices. The rest of the signal handlers
still need to be moved.