fix(windows): failing to connect to le device
Windows knew of the existence of the Bluetooth device, since the device
is already connected. It didn't, however, know that it is a Bluetooth
Low Energy device. We need to scan for LE devices to ensure Windows
knows of its LE capabilities.

Without doing this, the device will still show up in the list, but
connecting will fail with HRESULT 0 (meaning null, essentially).