fix(gui): actions/set_device error message handling
Since that action spawned a new async task, it wasn't returning any
errors that occur in that task. They were just directly being logged
with tracing.

Since we have to give away ownership of the task handle so it can be
canceled, we can't use the task's return value for the Result. A oneshot
channel is used instead. The task may be canceled, which will delete
the oneshot sender, causing the oneshot receiver to error. Nothing needs
to be done then, just stop waiting for the canceled task.