feat: increase equalizer range to -12db to +12db
The official soundcore app limits the range to -6db to +6db, but values
outside of that range seem to work fine. The 100hz range is notoriously
loud on the Q30s, so the extra -6db will be useful in bringing that
down.

The volume byte sent is centered with 0db being 0x120, so we can't go
any lower than -12db (0x00).