fix(web): device not being detected
In order for the device to be detected, the services first had to be
discovered by going to about://bluetooth-internals/#devices, scanning,
and inspecting the device.

We can filter the scan using manufacturerData instead to get around this
issue, and this way it's consistent with the desktop and android
versions that also filter by mac address.

The manufacturer data is in a non standard  format: mac address followed
by 0x00 0x00. So we can easily filter by first three bytes of the mac
address this way.