feat(android): add quick presets
This adds an option to show a notification with buttons for predefined
ambient sound mode, noise canceling mode, or equalizer profiles.

The number of quick presets is limited to two due to the limit of three
buttons on a notification. If more than two quick presets are needed in
the future, a custom notification layout may be necessary.