fix(lib): state update packet parsers should be all_consuming
It is possible that the state update packets would be distinguished by
something else failing to parse, such as an out of range sound mode, but
I intended for them to be distinguished by length. Wrapping the parser
in all_consuming to force it to fail if there is any remaining input
will handle this.