feat(lib): add partial support for DRC
Adds SetEqualizerWithDrcPacket and tracking of whether devices support
Dynamic Range Compression. The A3951 and A3930 support DRC regardless of
firmware version, so those should work as of this commit.

The other devices require firmware 02.00 or newer for DRC, and their
firmware version is not included in their state update packet. Since we
don't yet request/receive the firmware version, DRC support is not yet
enabled for those devices.