build(lib): remove tokio_rt_multi_thread wrapper crate
With dependency resolver v1, we couldn't have different features
enabled for dependencies depending on platform, so I created a wrapper
crate for tokio with rt-multi-thread enabled. I was unaware at the time
that resolver v1 was being used, since v2 is supposed to be the default
in edition 2021+, but apparently that does not apply when workspaces are
being used.

Resolver v2 was enabled in b121f7e1450c4c12237beaea7940ba165ffd88c3.