refactor: use float64 rather than i8 for for VolumeAdjustments
The range needed to be expanded from -120 to +120 to -120 to +135, so
the options were using i16 or larger, or using floats. i16 would make
some things simpler (eq comparisons, hashing, ordering, etc), but again
limits the precision. Using floats allows for varying the level of
precision displayed in the UI without having to make any lib changes.