feat(android): preset profiles specific to device type
It is unlikely that the same EQ profiles would be useful across devices,
so it makes sense to have quick presets be specific to a device rather
than shared across devices.

For backwards compatibility, if a preset is not set for a given device,
it will fall back to any shared quick presets that were created prior to
this commit.

In the future, maybe a way of clearing the old shared presets should be
added.