fix(android): disconnecting when rotating screen
If notifications are disabled, stopServiceIfNotificationIsGone will
always stop the service immediately. The ON_DESTROY event fires when the
screen is rotated, so that would stop the service. The view model does
not get destroyed when rotating the screen, so the service connection
should be bound to the view model lifecycle instead.