refactor(android): rename device_quick_preset table to match dao
The name was different originally since I wanted to rename the
old quick_preset table to fallback_quick_preset and then create a new
table by the name of quick_preset, but that seemed a bit tricky to do
using auto migrations, so I just chose a different name. Using a second
auto migration to rename the newly created table simplifies this.