fix(lib): use default values when sound mode is invalid
When the ambient sound mode, transparency mode, or noise canceling mode
is set to an invalid value, the device itself will remain on its last
valid value. In state update packets, however, it will report the
invalid value, so there is no way of knowing what the actual value is.

I decided it would be more appropriate to use a default value to get the
device into a good state as soon as possible rather than pointlessly
remembering an invalid value.