fix(android): coroutines started more than once
There should only be a single, persistent instance of some of
DeviceService's coroutines that were being started in the onStartCommand
method. That method is called every time startService is called, even if
the service is already running. This would start multiple instances.

The coroutines should be started in onCreate instead.