build: replace cargo-make with Just
cargo-make was originally chosen so that bash would not be a requirement
for building, but now that msys2 is required on windows, bash will
always be present.

Just is a better choice here since rust is not the only language used.

The install scripts are also now independent bash scripts that are not
dependent on Just, so they can be used when building the flatpak.