fix(android): custom equalizer profile duplicate detection
The equalizer profile volume adjustments were stored in sqlite as REALs,
aka floating point numbers, so equality comparisons for the unique index
won't work as expected.

Ideally, volume adjustments would be stored as DECIMAL(2, 1) or
something, but that doesn't exist in sqlite. Instead, they will be
stored as INTEGERs, but the values will be multiplied/divided by 10 when
inserting and retreiving.