fix(gui): write config to temp file first to ensure no data loss
If the program were to be killed while writing data to the file, or
perhaps other situations such as the drive being full, data could be
lost. The solution to this is to first write to a temporary file in the
same directory, and then `mv` the temp file to the original, which is
atomic.