test(gui): don't keep settings file open
Because NamedTempFile was used, the file is kept open from its creation
until it is deleted. The intention is for it to only be opened for
reading/writing and then immediately closed again. Not having that
behaviour causes issues on Windows, where you can't overwrite an open
file.

By using a temp dir instead, we can manage the files inside it
ourselves.