refactor(lib): merge parsing into the same files as structs
I originally decided to separate them since they are technically
separate things. There could be multiple parsers for the same struct. In
practice, though, that hasn't happened, and it's just an annoyance to
have them in separate files.

There was some dead code that didn't previously produce a warning since
it was pub rather than pub(crate). For now, it's marked with