fix(gui): crash on startup due to unexpected notify
Similar to another case of this same issue in the same file,
connect_selected_item_notify's handler was firing before we got a change
to set `sender`, but the handler expects it to be set, so it would
crash. We can ignore the case where it isn't set, since it will only
occur before connecting to a device.