docs: update android build instructions
ANDROID_NDK_HOME is no longer used when stripping debug info from
binaries. It still needs to be set for cargo-ndk, though.

Gradle build scripts were migrated to kotlin a long time ago, so the
file paths to the build scripts should end in .kts

If symlinks in toolchains/llvm/prebuilt/linux-x86_64/bin/ are not
preserved, you'll get an execve error when running cargo ndk. I'm adding
a note about it since it is a non trivial error message to diagnose.