fix(lib): A3936 state update packet parsing
I didn't have any test data to work with when I wrote the parser, so
I had no way to catch my mistakes. Now that I added a test, those issues
were revealed.

One remaining problem is the 0x49 button action. We only know what 0
through 6 are, so for now, the 9 is logged a default value of 0 is used
instead.