fix(lib): incorrect custom button action implementation
Each button is represented as a boolean "switch", followed by a byte
with the left four bits representing the action that was ASSIGNED while
tws was disconnected, and the right four being the action that was
assigned when tws was connected. The switch is then set to 0 or 1 to
select which of those two actions should be used.

For single presses, there is nothing in the left four bits, and the
switch is actually just enabled/disabled.