fix(gui): quick presets unique by model rather than gatt service id
Since BLE is no longer used, the gatt service should no longer be used
to identify devices. The first four characters of the serial number seem
to be the model number, so we can use that to identify device types.

Old presets will be automatically migrated upon connection. In the
future (at least 6 months from now), once most people have their
presets migrated over, v2.0 can be released and remove the migration
code.

If there was no downside, the preset migration would be kept around
forever, but having it around means that devices that don't support BLE
will be slower to connect, since it has to wait for the BLE connection
to time out.