fix(lib): windows connection issues with multiple bluetooth adapters
I don't know for sure that this is what is going on, but my
understanding is that BluetoothDevice::FromBluetoothDeviceAsync is
returning the device from the inactive bluetooth adapter rather than the
active one. As a fix, we can enumerate all devices and filter for only
connected ones with the correct mac address.