fix(gui): windows installer not installing older versions
The installer currently compares version numbers of all files being
installed, and doesn't install them if the existing file is newer or
the same version. No indication of this is given to the user, so I'm
disabling that functionality. Files will be replaced no matter what the
version is. I don't intend to go back and generate new installers for
all existing old versions though, so those installers will still have
this problem.