lib!: snapshot settings at quick preset creation
Rather than using the value of the setting when the user enables it for
the quick preset, all settings are now snapshotted upon creation of the
quick preset. This way toggling a field off and back on is no longer a
destructive operation.

This is a breaking change, since it modifies the database format with no
migration path.