refactor: reuse error types less
This allows for more specific error messages. There were also mistakes
where errors would both have a source and include the source's Display
output in their Display output, when they should only do one or the
other.

I'm not satisfied with how this commit leaves things, but it's an
improvement.