fix(lib): race condition when changing settings right after connecting
If you change a setting before the initial StateUpdatePacket is
processed, the change can get reverted by that. The current fix is a
hack, and it would be better to instead have an option in
PacketIOController to only return the response and don't forward it to
the listeners for further processing.