fix(lib): incorrect a3959 button action parsing
Most other devices have one byte for whether the button is
enabled/disabled, and a second byte for the action. This device uses a
single byte that is either the action or 0xF if disabled. This was
identified correctly when documenting the packet in
tools/soundcore-device-faker, but not implemented in openscq30-lib.

The triple press action was parsed, but was not exposed as a setting.
This adds setting ids for left and right triple press.

GH-149