refactor(lib): add button_configuration_v2
This implements a configurable button configuration module so that
devices with varying button configurations can define those via
configuration rather than implementing a new module for that specific
device.

It makes the implementation of the module more messy, but I want to
prioritize reducing friction in adding support for new devices.