fix(lib): incorrect a3948 button configuration handling
The a3948 has an enabled flag for each button, but it does not make use
of it. Instead, the action is set to 0xF to disable. I'm not sure if the
device ignores the disabled flag or not, but to be safe, if the button
got disabled somehow (older version of openscq30?), it will be enabled
again.