feat: reset buttons to default
Adds a setting for resetting buttons to their default actions.

In order to prevent the button_configuration module from immediately
reverting the changes made by the reset_button_configuration module, a
boolean is shared between them to track if buttons are being reset. In
that case, the button_configuration module can do nothing and return.