fix(lib): fix a3947 and a3948 respecting button enabled flag
Rather than requiring both the enabled flag and a non 0xF action to be
considered enabled, the enabled flag should be ignored entirely,
treating the button as enabled even if the flag is false.

This also renames the ButtonConfigurationSettings boolean so that false
is the sensible default value rather than true.