feat(lib)!: adjust format of battery level string
For the non-localized string, it will now include the total. So instead
of "2", it will be "2/5".

For the localized string, it is now a percentage. Instead of "2/5", it
will be "40%".

I'm considering this a breaking change since it changes a non-localized
string, which affects the CLI's output. That's okay, since v2.0.0 is not
yet released.