fix(android): foreground service only binds on first autoconnect
When the foreground service is started by something other than
OpenSCQ30RootViewModel, there was nothing to inform the viewmodel to
bind to the service.

This also removes some unnecessary code, since when I originally wrote
it, I thought the service wouldn't stop when calling stopservice if
anything was bound to it, but that's not the case.