deps: update cargo dependencies
sqlite doesn't actually support u64's (and therefore usize on 64 bit
platforms), so support for u64s was put behind a feature flag in
rusqlite. This affects the current migration number (stored in
pragma_user_version.user_version), which we were using usize for. It
could be changed to either i64 or u32. I went with u32 since it
disallows negative values and is well over the number of migrations we
will ever have.