build: provide stable commands for building
By providing `just build-*` commands in the project root that output
artifacts to `./build-output`, it is possible to make significant
changes to the build process without requiring packaging changes.

Of course, this doesn't apply to situations where `just` is inconvenient
to install (flatpak, windows ci), so large changes will still be
avoided.